requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.8
python3-xlib==0.15
pytweening==1.0.4
